research_stocks/pyproject.toml
[tool.poetry]
name = "research_stocks"
version = "0.1.0"
description = "research_stocks using CrewAI"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "research_stocks", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = { extras = ["tools"], version = ">=0.121.1,<1.0.0" }
python-dotenv = "^1.0.1"
html2text = "^2024.2.26"
sec-api = "^1.0.20"

[tool.poetry.scripts]
research_stocks = "research_stocks.main:run"
run_crew = "research_stocks.main:run"
train = "research_stocks.main:train"
replay = "research_stocks.main:replay"
test = "research_stocks.main:test"

[tool.poetry.group.dev.dependencies]
setuptools = ">=68"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.crewai]
type = "crew"